Move and adapt tests from sql/storedproc
