+++ hibernate-core/src/test/java/org/hibernate/orm/test/sql/storedproc/H2ProcTesting.java
@@ -4,9 +4,8 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html
  */
-package org.hibernate.test.sql.storedproc;
+package org.hibernate.orm.test.sql.storedproc;
 
-import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.EntityResult;
 import javax.persistence.FieldResult;
@@ -17,6 +16,7 @@ import javax.persistence.QueryHint;
 import javax.persistence.SqlResultSetMapping;
 import javax.persistence.StoredProcedureParameter;
 
+import org.hibernate.boot.MetadataBuilder;
 import org.hibernate.boot.model.relational.AuxiliaryDatabaseObject;
 import org.hibernate.cfg.Configuration;
 import org.hibernate.dialect.Dialect;
@@ -26,8 +26,8 @@ import org.hibernate.dialect.H2Dialect;
  * @author Steve Ebersole
  */
 public class H2ProcTesting {
-	public static void applyProcDefinitions(Configuration configuration) {
-		configuration.addAuxiliaryDatabaseObject(
+	public static void applyProcDefinitions(MetadataBuilder metadataBuilder) {
+		metadataBuilder.applyAuxiliaryDatabaseObject(
 				new AuxiliaryDatabaseObject() {
 					@Override
 					public String getExportIdentifier() {
@@ -71,7 +71,7 @@ public class H2ProcTesting {
 				}
 		);
 
-		configuration.addAuxiliaryDatabaseObject(
+		metadataBuilder.applyAuxiliaryDatabaseObject(
 				new AuxiliaryDatabaseObject() {
 					@Override
 					public String getExportIdentifier() {
@@ -115,7 +115,7 @@ public class H2ProcTesting {
 				}
 		);
 
-		configuration.addAuxiliaryDatabaseObject(
+		metadataBuilder.applyAuxiliaryDatabaseObject(
 				new AuxiliaryDatabaseObject() {
 					@Override
 					public String getExportIdentifier() {
