+++ hibernate-core/src/test_legacy/org/hibernate/test/sql/storedproc/StoredProcedureParameterRegistrationTest.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html
  */
-package org.hibernate.test.sql.storedproc;
+package org.hibernate.orm.test.sql.storedproc;
 
 import java.util.List;
 import javax.persistence.ParameterMode;
